refactor(reviewhub): decouple state from modal chrome
- Extract all state, IPC handlers, and effects into useReviewHubState hook
- Create ReviewHubBody for the reusable presentation surface
- Slim ReviewHub to modal chrome (portal, overlay, keyboard, backdrop)
- Hook accepts {worktreePath, active} so non-modal containers can drive it
- Preserve combined init/reset effect and request-id cancellation pattern
